+++ paleoDiv/R/paleoDiv.R
@@ -196,7 +196,7 @@ return(y_)
 #'
 #' @param x Variable for which to plot violin.
 #' @param pos Position at which to place violin in the axis perpendicular to x. Defaults to 0
-#' @param x2 Optional variable to use instead of x as input variable for the violin plot. If x2 is set, the function (default: density()) used to calculate the plotting statistic is run on x2 instead of x, but the results are plotted at the corresponding x values.
+#' @param x2 Optional variable to override the use of x as input variable for the plotting statistic. If x2 is set, the function (default: density()) used to calculate the plotting statistic is run on x2 instead of x, but the results are plotted at the corresponding values of x.
 #' @param stat The plotting statistic. Details to the density() function, as in a standard violin plot, but can be overridden with another function that can take x or x2 as its first argument. Stat can also be a numeric vector of the same length as x, in which case the values in this vectors are used instead of the function output and plotted against x as an independent variable.
 #' @param dscale The scale to apply to the values for density (or another plotting statistic). Defaults to 1, but adjustment may be needed depending on the scale of the plot the violin is to be added to.
 #' @param cutoff Setting for cropping the violin. Can be either a single value, in which case the input is interpreted as number of standard deviations from the mean, or a numeric vector of length 2, giving the lower and upper cutoff value directly.
@@ -209,6 +209,7 @@ return(y_)
 #' @param col Line color for the plotted violin
 #' @param lwd Line width for the plotted violin
 #' @param lty Line width for the plotted violin
+#' @param na.rm logical indicating whether to remove NA values from input data.
 #' @param ... Other arguments to be passed on to function in parameter stat
 #' @return A violin plot and a data.frame containing the original and modified plotting statistic and independent variable against which it is plotted.
 #' @details
@@ -221,8 +222,18 @@ return(y_)
 #' viol(c(1:10), width=9, stat=rmean, pos=0, add=FALSE)
 #' viol(c(1:10), stat=c(11:20), pos=0, add=FALSE)
 
-viol<-function(x, pos=0, x2=NULL, stat=density, dscale=1, cutoff=range(x), horiz=TRUE, add=TRUE,lim=cutoff,xlab="",ylab="", fill="grey", col="black", lwd=1, lty=1,...){
+viol<-function(x, pos=0, x2=NULL, stat=density, dscale=1, cutoff=range(x), horiz=TRUE, add=TRUE,lim=cutoff,xlab="",ylab="", fill="grey", col="black", lwd=1, lty=1, na.rm=FALSE,...){
 #sort ascending
+
+if(na.rm){#remove NAs
+which(!is.na(x))->notNA
+if(!is.null(x2)) intersect(notNA,which(!is.na(x2)))->notNA
+
+x[notNA]->x
+if(!is.null(x2)) x2[notNA]->x2
+}
+
+
 if(!is.null(x2) & is.numeric(x2) & length(x2)==length(x)){
 x2[order(x)]->x2
 }
@@ -283,6 +294,298 @@ invisible(data.frame(o_statistic=c(dstat0_,rev(dstat0_)),plot_statistic=c(pos+ds
 
 }
 ##
+
+
+##function violins()
+#' Wrapper around viol() to conveniently plot multiple violins on a single plot, analogous to the behavior of boxplot()
+#' @param x plotting statistic (numeric vector) or formula object from which a plotting statistic and grouping variable can be extracted (i.e. of form x~group)
+#' @param group grouping variable
+#' @param horiz logical indicating whether to plot horizontally
+#' @param order order of factor levels of categorical factor
+#' @param data data.frame object containing x and y
+#' @param xlab x axis label
+#' @param ylab y axis label
+#' @param col vector of border colors
+#' @param fill vector of fill colors
+#' @param lwd vector of line widths
+#' @param lty vector of line types
+#' @param xlim x limits (data limits used if NULL)
+#' @param ylim y limits (data limits used if NULL)
+#' @param spaces character string in group to replace with spaces for labels, if not NULL
+#' @param dscale density scaling factors (numeric) to apply to individual violins
+#' @param add logical whether to add to existing plot (default: FALSE)
+#' @param ax whether to plot axes
+#' @param srt angle for categorical axis text rotation
+#' @param ... other arguments to pass on to paleoDiv::viol() and plot()
+#' @export violins
+#' @importFrom paleoDiv viol
+#' @examples 
+#' data.frame(p=rnorm(50), cat=rep(c("A","B","B","B","B"),10))->d
+#' violins(p~cat,d)
+
+violins<-function(x, data=NULL, group=NULL, horiz=FALSE, order=NULL, xlab="", ylab="", col="black",fill="grey", lwd=1, lty=1,dscale=1,xlim=NULL, ylim=NULL, spaces="_", add=FALSE, ax=TRUE,srt=45,...){
+
+if(ax){
+pr<-function(axis="x"){#helper function plotr for label plotting
+if(axis=="x"){
+abs(diff(range(par("usr")[1:2])))
+}else{
+abs(diff(range(par("usr")[3:4])))
+}
+}}
+
+if(inherits(x,"formula")){
+
+if(x[1]!=`~`()) stop("no ~ operator found in formula supplied to x")
+
+if(!is.null(data)){
+data[[as.character(x[2])]]->x_
+data[[as.character(x[3])]]->group
+x_->x
+}else{
+get(as.character(x[2]))->x_
+get(as.character(x[3]))->group
+x_->x
+}
+#print(data.frame(x=x,group=group))
+
+}
+
+if(!is.null(spaces)){
+gsub(spaces," ", group)->group
+}
+
+
+##xrange and number of categories
+range(x)->rx
+levels(factor(group))->cat
+length(cat)->ncat
+
+if(!is.null(order)) cat[order]->cat
+
+##visual settings
+if(length(col)<ncat) rep(col,ncat)[1:ncat]->col
+if(length(fill)<ncat) rep(fill,ncat)[1:ncat]->fill
+if(length(lwd)<ncat) rep(lwd,ncat)[1:ncat]->lwd
+if(length(lty)<ncat) rep(lty,ncat)[1:ncat]->lty
+##XXX##
+if(length(dscale)<ncat) rep(dscale,ncat)[1:ncat]->dscale
+
+##conditional plot limits
+if(horiz){
+if(is.null(xlim)) xlim<-rx
+if(is.null(ylim)) ylim<-c(0,ncat+1)
+}else{
+if(is.null(xlim)) xlim<-c(0,ncat+1)
+if(is.null(ylim)) ylim<-rx
+}
+
+
+##now plot
+if(add==FALSE) plot(NA,type="n", axes=F, ylim=ylim, xlim=xlim,xlab=xlab, ylab=ylab,...)#base plot
+
+##add viols
+if(horiz==T){#horizontal viols
+for(i in 1:ncat){#loop
+paleoDiv::viol(x=x[group==cat[i]], pos=i, horiz=TRUE, fill=fill[i], col=col[i], lwd=lwd[i], lty=lty[i],dscale=dscale[i],...)
+}#end loop
+
+if(ax){
+axis(1)
+#mtext(side=2, at=c(1:ncat), text=cat, col=col)
+text(x=par("usr")[1]-pr("x")*0.015,xpd=T, srt=srt, adj=c(1,0), y=c(1:ncat), col=col, cat)
+
+}
+
+}else{#vertical viols
+for(i in 1:ncat){#loop
+paleoDiv::viol(x=x[group==cat[i]], pos=i, horiz=FALSE, fill=fill[i], col=col[i], lwd=lwd[i], lty=lty[i],dscale=dscale[i],...)
+}#end loop
+
+if(ax){
+axis(2)
+
+#mtext(side=1, at=c(1:ncat), text=cat, col=col)
+text(y=par("usr")[3]-pr("y")*0.015,xpd=T, srt=srt, adj=c(1,0), x=c(1:ncat), col=col, cat)
+
+}
+
+}
+
+
+out<-1:ncat
+names(out)<-cat
+invisible(out)
+
+}
+##
+
+
+
+##function jitterp()
+#' plot data as a jitter-plot
+#' @param x x values to plot (if single value and y is a vector, plot is vertical)
+#' @param y y value at which to plot (if single value, plot is horizontal)
+#' @param width standard deviation for jitter
+#' @param col color for points
+#' @param alpha opacity for points
+#' @param ... other parameters to be passed on to points()
+#' @return adds the points to the open plotting device as a jitter plot and returns an invisible list()-object containing the positions of all points
+#' @importFrom graphics points
+#' @importFrom stats rnorm
+#' @importFrom paleoDiv add.alpha
+#' @export jitterp
+#' @examples
+#' c(1,2,3,2,3,2,3,4,4)->tmp
+#' hist(tmp)
+#' jitterp(x=tmp, y=1, width=0.1)
+
+jitterp<-function(x,y,width,col="black",alpha=0.5,...){
+
+if(length(y)==1){
+y<-rep(y,length(x))+rnorm(n=length(x), mean=0, sd=width)
+}else if(length(x)==1){
+x<-rep(x,length(y))+rnorm(n=length(y), mean=0, sd=width)
+}else{
+stop("Please make sure either x or y is length()==1")
+}
+
+points(x=x, y=y,col=paleoDiv::add.alpha(col,alpha),...)
+
+invisible(data.frame(x,y))
+
+}
+##
+
+
+##function multijitter()
+#' Wrapper around jitterp that plots multiple jitter plots on the same plotting device (analogous to violins())
+#' @param x plotting statistic (numeric vector) or formula object from which a plotting statistic and grouping variable can be extracter (i.e. of form x~group)
+#' @param group grouping variable
+#' @param horiz logical indicating whether to plot horizontally
+#' @param order order of factor levels of categorical factor
+#' @param data data.frame object containing x and y
+#' @param xlab x axis label
+#' @param ylab y axis label
+#' @param col vector of border colors
+#' @param pch vector of symbols
+#' @param xlim x limits (data limits used if NULL)
+#' @param ylim y limits (data limits used if NULL)
+#' @param spaces character string in group to replace with spaces for labels, if not NULL
+#' @param width standard deviation for jitter
+#' @param ax whether to plot axes
+#' @param srt angle for categorical axis text rotation
+#' @param add logical whether to add to existing plot (default: TRUE)
+#' @param ... other arguments to pass on to jitterp() and plot()
+#' @importFrom graphics axis
+#' @importFrom graphics mtext
+#' @export multijitter
+#' @examples 
+#' data.frame(p=rnorm(50), cat=rep(c("A","B","B","B","B"),10))->d
+#' multijitter(p~cat,d, add=FALSE)
+
+multijitter<-function(x, data=NULL, group=NULL, horiz=FALSE, order=NULL, xlab="", ylab="", col="black", pch=16, spaces="_", width=0.1, xlim=NULL, ylim=NULL,add=TRUE,ax=FALSE,srt=45,...){
+if(ax){
+pr<-function(axis="x"){#helper function plotr for label plotting
+if(axis=="x"){
+abs(diff(range(par("usr")[1:2])))
+}else{
+abs(diff(range(par("usr")[3:4])))
+}
+}}
+
+
+if(inherits(x,"formula")){
+
+if(x[1]!=`~`()) stop("no ~ operator found in formula supplied to x")
+
+if(!is.null(data)){
+data[[as.character(x[2])]]->x_
+data[[as.character(x[3])]]->group
+x_->x
+}else{
+get(as.character(x[2]))->x_
+get(as.character(x[3]))->group
+x_->x
+}
+#print(data.frame(x=x,group=group))
+
+}
+
+
+if(!is.null(spaces)){
+gsub(spaces," ", group)->group
+}
+
+##xrange and number of categories
+range(x)->rx
+levels(factor(group))->cat
+length(cat)->ncat
+
+if(!is.null(order)) cat[order]->cat
+
+##visual settings
+if(length(col)<ncat) rep(col,ncat)[1:ncat]->col
+if(length(pch)<ncat) rep(pch,ncat)[1:ncat]->pch
+
+##conditional plot limits
+if(horiz){
+if(is.null(xlim)) xlim<-rx
+if(is.null(ylim)) ylim<-c(0,ncat+1)
+}else{
+if(is.null(xlim)) xlim<-c(0,ncat+1)
+if(is.null(ylim)) ylim<-rx
+}
+
+
+##now plot
+if(add==FALSE) plot(NA,type="n", axes=F, ylim=ylim, xlim=xlim,xlab=xlab, ylab=ylab,...)#base plot
+
+out<-list()
+
+##add jitterplots
+if(horiz==T){#horizontal viols
+for(i in 1:ncat){#loop
+
+
+jitterp(x=x[group==cat[i]], y=i, width=width, col=col[i], pch=pch[i],...)->out[[i]]
+
+
+}#end loop
+
+if(ax){
+axis(1)
+#mtext(side=2, at=c(1:ncat), text=cat, col=col)
+text(x=par("usr")[1]-pr("x")*0.015,xpd=T, srt=srt, adj=c(1,0), y=c(1:ncat), col=col, cat)
+
+}
+
+}else{#vertical viols
+for(i in 1:ncat){#loop
+
+
+jitterp(y=x[group==cat[i]], x=i, width=width, col=col[i], pch=pch[i],...)->out[[i]]
+
+
+}#end loop
+
+if(ax){
+axis(2)
+
+#mtext(side=1, at=c(1:ncat), text=cat, col=col)
+text(y=par("usr")[3]-pr("y")*0.015,xpd=T, srt=srt, adj=c(1,0), x=c(1:ncat), col=col, cat)
+
+}
+
+}
+
+invisible(out)
+
+}
+##
+
+
+
 
 ##Function tsconv
 #'Convert geological ages for accurate plotting alongside a calibrated phylogeny
@@ -492,7 +795,7 @@ c(min(lastPP),min(lastPP)+ylim)->ylim
 #' @param arrow.angle Angle of arrow ends to be used for plotting. Defaults to 45 degrees.
 #' @param arrow.code Arrow code to be used for plotting. For details, see ?arrows
 #' @param indices Optional indices which edges to redraw. Can be used to highlight specific edges in different color or style.
-#' @return Plots a timescale on the currently active plot.
+#' @return Nothing (redraws selected edges of the phylogeny on the active plot device)
 #' @importFrom graphics arrows
 #' @importFrom ape plot.phylo
 #' @export redraw.phylo
@@ -541,31 +844,48 @@ stop("Only phylogram and cladogram supported so far")
 #' @param full A logical indicating whether or not the full dataset is to be downloaded (defaults to FALSE). At the expense of larger file size, the full dataset contains a large number of additional columns containing data such as stratigraphy, phylogeny and (paleo)geography, which is useful for various purposes but not strictly necessary for graphing paleodiversity.
 #' @param base Character string containing base url to use. Defaults to https://paleobiodb.org/data1.2/. Entering "dev" serves as a shortcut to use https://dev.paleobiodb.org/data1.2/ instead (can sometimes be helpful if one of the two is unavailable).
 #' @param file Character string containing which file name to look for. Defaults to list.csv.
+#' @param cc Selection for continent (e.g. EUR for Europe, see paleobiodb.org documentation)
+#' @param envtype Selection for environment type (e.g. marine)
+#' @param append_additional Any additional character string to append to URL for pdb dataset
 #' @return A data.frame() containing the downloaded paleobioDB dataset. The column "identified_name" will be copied into the column "tna", and (if what==occs) the columns "max_ma" and "min_ma" will be copied into the columns named "eag" and "lag" respectively, maintaining compatibility with the output of the deprecated package "paleobioDB" for those variable names.
 #' @importFrom utils read.csv
 #' @export pdb
 #' @examples
 #' pdb("Stegosauria")->Stegosauria
 
-pdb<-function(taxon, interval="all", what="occs", full=FALSE, base="https://paleobiodb.org/data1.2/",file="list.csv"){
+pdb<-function(taxon, interval="all", what="occs", full=FALSE, base="https://paleobiodb.org/data1.2/",file="list.csv", cc=NULL, envtype=NULL, append_additional=NULL){
 
 if(base=="dev"){base<-"https://dev.paleobiodb.org/data1.2/"}
 
-if(full==TRUE){
-    if(interval=="all"){
-    pbdb_url <-paste0(base,what,"/",file,"?base_name=",taxon,"&show=full")
-    }else{
-    pbdb_url <-paste0(base,what,"/",file,"?base_name=",taxon,"&interval=",interval,"&show=full")}
-}else{
-    if(interval=="all"){
+##generate occurrence url
+if(interval=="all"){
     pbdb_url <-paste0(base,what,"/",file,"?base_name=",taxon)
     }else{
     pbdb_url <-paste0(base,what,"/",file,"?base_name=",taxon,"&interval=",interval)}
+
+	if(full==TRUE){#specify to download full data record if full==TRUE
+    pbdb_url <-paste0(pbdb_url,"&show=full")
     }
 
+    if(!is.null(cc)){#add continent selection, if stated
+		if(is.character(cc)){
+		pbdb_url <-paste0(pbdb_url,"&cc=",cc)
+		}
+    }
+    
+    if(!is.null(envtype)){#add envtype selection, if stated
+		if(is.character(envtype)){
+		pbdb_url <-paste0(pbdb_url,"&envtype=",envtype)
+		}
+    }
+    
+     if(!is.null(append_additional)){#add additional string to append to URL, if stated
+		if(is.character(append_additional)){
+		pbdb_url <-paste0(pbdb_url,append_additional)
+		}
+    }
 
-#occ<-read.csv(pbdb_url)
-
+##now read csv from generated url
 tryCatch({read.csv(pbdb_url)}, error=function(e){
 return(paste("error in paleodb query:", e$message))
 
@@ -575,14 +895,15 @@ return(paste("warning in paleodb query:", w$message))
 })->occ
 
 
+#copy some columns
 if(is.data.frame(occ)){
     if(what=="occs"){
 occ$identified_name->occ$tna#for colname compatibility with deprecated paleobiodb package
 }
 
-occ$hltax<-taxon#higher level taxon for categorization,e.g. if tables are combined via rbind
-occ$min_ma->occ$lag#for colname compatibility with deprecated paleobiodb package
-occ$max_ma->occ$eag#for colname compatibility with deprecated paleobiodb package
+occ$hltax<-taxon#higher level taxon for categorization, e.g. if tables are combined via rbind
+occ$min_ma->occ$lag#for compatibility with deprecated paleobiodb package
+occ$max_ma->occ$eag#for compatibility with deprecated paleobiodb package
 return(occ)
 
 }else if(length(occ)==1){
@@ -959,12 +1280,13 @@ return(length)}#standard setting ids=FALSE returns number of records
 #' @param taxa Either a character vector of valid taxonomic names, or an object of class "phylo" whose tip.labels to use instead.
 #' @param cleanup Logical indicating whether to apply occ.cleanup() to occurrence data after download (defaults to TRUE)
 #' @param interval Stratigraphic interval for which to download data (defaults to NULL, which downloads data for all intervals)
+#' @param ... additional arguments to be passed on to pdb()
 #' @return A list() object containing occurrence data (saved under the taxon names given) and species-level taxon-range tables (saved with the prefix "sptab_" before the taxon names).
 #' @export pdb.autodiv
 #' @examples
 #' pdb.autodiv("Coelophysoidea")->coelo
 
-pdb.autodiv<-function(taxa,cleanup=TRUE,interval=NULL){
+pdb.autodiv<-function(taxa,cleanup=TRUE,interval=NULL,...){
 occ<-list()
 if(inherits(taxa, "phylo")){
 taxa$tip.label->treetips
@@ -974,9 +1296,9 @@ taxa$tip.label->treetips
 for(i in 1:length(treetips)){
 
 if(is.null(interval)){
-pdb(treetips[i])->occ[[i]]
+pdb(treetips[i],...)->occ[[i]]
 }else{
-pdb(treetips[i],interval=interval)->occ[[i]]
+pdb(treetips[i],interval=interval,...)->occ[[i]]
 }
 
 
@@ -1077,7 +1399,7 @@ colnames(ages)<-c("FAD", "LAD")
 rownames(ages)<-taxa
 
 if(e==1){
-warning("Some occurrence tables could not be found, corresponding collumns contain NAs.")
+warning("Some occurrence tables could not be found, corresponding columns contain NAs.")
 }
 
 return(ages)
@@ -1090,7 +1412,7 @@ return(ages)
 #' @param phylo0 Either an object of class phylo, or a character vector containing taxon names for building the matrix
 #' @param data A higher-level taxon name to get data for in the paleobiology database, or a data.frame containing a species table containing entries for the taxa in question.
 #' @return A two-column matrix containing earliest and latest occurrences for each taxon in taxa, with taxon names as row names
-#' @details tree.ages looks for the taxon names in the tna collumn of a taxon-range table (as produced by mk.sptab()), so it will only recover ages for taxa that can be found there. For a function optimized for higher-level taxa that might not be represented in such a table, see tree.ages().  It is highly recommended to manually inspect the resulting table for accuracy.
+#' @details tree.ages looks for the taxon names in the tna column of a taxon-range table (as produced by mk.sptab()), so it will only recover ages for taxa that can be found there. For a function optimized for higher-level taxa that might not be represented in such a table, see tree.ages().  It is highly recommended to manually inspect the resulting table for accuracy.
 #' @importFrom utils read.csv
 #' @export tree.ages.spp
 #' @examples
@@ -1406,9 +1728,9 @@ plot1<-tryCatch({get("last_plot.phylo", envir = ape::.PlotPhyloEnv)}, error=func
 
 }
 
-if(is.null(plot1)){
-plot1$x.lim<-xlimits
-}
+#changed in v. 0.3.5 to use actual plot margins
+#if(is.null(plot1)){plot1$x.lim<-xlimits}
+plot1$x.lim<-par("usr")[1:2]
 
 #repeat colors into vectors, if needed
 col->col_